+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [1.0.0-alpha.3] - 2019-10-05
+### Fixed
+- anno_helper.divide_into_patches: output of shifted bbox coordinates.
+- blacklist: fixed behaviour of get_whitelist method. Do not put images automatically on blacklist.
+
+
 ## [1.0.0-alpha.2] - 2019-10-04
 ### Added
 - SIA: Added button to delete all 2D annos